+++ docs/changelog.rst
@@ -9,7 +9,7 @@ Changelog
 Added
 ~~~~~
 
-- support for property-level examples. `#637`_
+- Support for property-level examples. `#467`_
 
 `2.0.0`_ - 2020-07-01
 ---------------------
@@ -1209,7 +1209,6 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
-.. _#467: https://github.com/kiwicom/schemathesis/issues/467
 .. _#631: https://github.com/kiwicom/schemathesis/issues/631
 .. _#629: https://github.com/kiwicom/schemathesis/issues/629
 .. _#621: https://github.com/kiwicom/schemathesis/issues/621
@@ -1254,6 +1253,7 @@ Fixed
 .. _#473: https://github.com/kiwicom/schemathesis/issues/473
 .. _#469: https://github.com/kiwicom/schemathesis/issues/469
 .. _#468: https://github.com/kiwicom/schemathesis/issues/468
+.. _#467: https://github.com/kiwicom/schemathesis/issues/467
 .. _#463: https://github.com/kiwicom/schemathesis/issues/463
 .. _#461: https://github.com/kiwicom/schemathesis/issues/461
 .. _#458: https://github.com/kiwicom/schemathesis/issues/458